Add option "wait"
This would be implemented by someone running `$(".equal-heights").equalHeights({ wait: true });`
Basically, it keeps running equalHeights until the maxHeight is greater than 0. It's really annoying when equalHeights is implemented at a time before the content has loaded and it sets the elements to be a height of 0px. This prevents this only when specifically requested by the user. Also, this opens the door to have other potential options in the future.